wip: fallback builder refactor
